+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 
+## [Unreleased] - 2023-04-30
+### Add
+- Add `onlyFirstLevel` method to the Repository class to return only the first level of the records.
+
 ## [0.0.0] - 2023-04-27
 ### Add
 - [af51c0d](https://github.com/Briofy/repository-laravel/commit/af51c0d678fe25dee24425848537c6903bc01d49) Initial commit of the project